Cart javascript

+++ app/Http/Controllers/Shop/CartController.php
@@ -2,8 +2,6 @@
 
 namespace App\Http\Controllers\Shop;
 
-use App\Models\Shop\Order;
-use App\Models\Shop\Product;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 
@@ -24,8 +22,9 @@ class CartController extends BaseController
     public function index()
     {
         $orderId = session()->get('orderId');
-        $this->data['order'] = $this->orderRepository->find($orderId);
-
+        $order = $this->orderRepository->find($orderId);
+        $this->data['order'] = $order;
+        $this->data['cartCount'] = $order->cartCount();
         return view('shop.userCart', $this->data);
     }
 
@@ -45,30 +44,14 @@ class CartController extends BaseController
         } else {
             $order = $this->orderRepository->find($orderId);
         }
-
-        if ($order->products->contains($product)) { // if isset
-            $pivotRow = $order->products()->find('product_id', $product->id)->first()->pivot;
-            $pivotRow->count++;
-            $pivotRow->update();
+        if ($order->products->contains($product)) {
+            $inc = $request->input('inc');
+            $this->orderRepository->pivotCount($product, $inc);
         } else {
             $order->products()->attach($product); // соединить модели pivot
         }
 
-//        $cart = session()->has('cart') ? session()->get('cart') : [];
-//        if (array_key_exists($product->id, $cart)) {
-//            $cart[$product->id]['quantity'] = $request->input('qty');
-//        } else {
-//            $order = Order::create();
-//            $cart[$product->id] = [
-//                'id' => $product->id,
-//                'order_id' => $order->id,
-//                'quantity' => $request->input('qty')
-//            ];
-//        }
-//        session(['cart' => $cart]);
-//        session()->flash('message', $product->title.' added to cart.');
-
-        $this->data['order'] = $order;
+        $this->data['cartCount'] = $order->cartCount();
         $this->data['message'] = $product->title . ' added to cart.';
         return response()->json($this->data);
     }
@@ -81,8 +64,14 @@ class CartController extends BaseController
     {
         $orderId = session()->get('orderId');
         $order = $this->orderRepository->find($orderId);
-        $order->products()->detach($product);
+
+        if ($this->productRepository->getProductFirst($product)) {
+            $order->products()->detach($product);
+        }
+
         $this->data['message'] = $product->title . ' remove from cart.';
         return response()->json($this->data);
     }
+
+
 }
